BaseButton component moved to src/components/base/
