+++ src/App.vue
@@ -1,7 +1,7 @@
 <script>
 import { defineComponent } from 'vue';
 
-import BaseButton from '@/components/BaseButton.vue';
+import BaseButton from '@/components/base/BaseButton.vue';
 
 export default defineComponent({
   components: {
